+++ packages/cli/src/cli-hosting.js
@@ -61,6 +61,7 @@ const setup = async () => {
     .command('sync <name>')
     .group('Hosting')
     .description('Publish your local hosting files')
+    .option('-d, --no-delete', 'Update and add files only (do not delete files)')
     .option('-s, --socket <name>', 'socket')
     .action(async (...options) => {
       session.isAuthenticated()
@@ -76,7 +77,7 @@ const setup = async () => {
     .description('Configure hosting parameters')
     .option('-c, --cname <domain_name>', 'add CNAME to hosting')
     .option('-d, --remove-cname <domain_name>', 'remove CNAME from hosting')
-    .option('-b, --browser_router', 'turn on BrowserRouter support')
+    .option('-b, --browser-router <true|false>', 'turn on/off the BrowserRouter support')
     .action(async (...options) => {
       session.isAuthenticated()
       session.hasProject()