+++ packages/lib-js-core/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.9.3"></a>
+## [0.9.3](https://github.com/Syncano/syncano-node/compare/v0.9.1...v0.9.3) (2018-02-19)
+
+
+
+
+**Note:** Version bump only for package @syncano/core
+
+<a name="0.9.0"></a>
+# [0.9.0](https://github.com/Syncano/syncano-node/compare/v0.8.0...v0.9.0) (2018-02-19)
+
+
+
+
+**Note:** Version bump only for package @syncano/core
+
 <a name="0.8.0"></a>
 # [0.8.0](https://github.com/Syncano/syncano-node/compare/v0.7.1...v0.8.0) (2018-02-09)